feat: enable settings revamp
