+++ apps/extension/src/app/components/layout/index.ts
@@ -4,5 +4,4 @@ export * from './card';
 export { AvailableBalance } from './card/components/available-balance';
 
 export { Content } from './layouts/content.layout';
-export { TwoColumnLayout } from './layouts/two-column.layout';
 export { ContainerLayout } from './layouts/container.layout';
