--- apps/extension/src/app/features/add-network/network-form.layout.tsx
@@ -1,20 +1,0 @@
-import { ReactNode } from 'react';
-
-import { Content, Page } from '@app/components/layout';
-import { PageHeader } from '@app/features/container/headers/page.header';
-
-interface NetworkFormLayoutProps {
-  title: string;
-  children: ReactNode;
-}
-
-export function NetworkFormLayout({ title, children }: NetworkFormLayoutProps) {
-  return (
-    <>
-      <PageHeader title={title} />
-      <Content>
-        <Page>{children}</Page>
-      </Content>
-    </>
-  );
-}
