+++ apps/extension/src/app/pages/view-secret-key/unlocked-view-secret-key.tsx
@@ -1,0 +1,35 @@
+import { Content } from '@app/components/layout';
+import { Header } from '@app/components/layout/headers/header';
+import { HeaderBackButton } from '@app/components/layout/headers/header-back-button';
+import { HeaderGrid } from '@app/components/layout/headers/header-grid';
+import {
+  DescriptionColumn,
+  TwoColumnLayout,
+} from '@app/components/layout/layouts/two-column.layout';
+import { useDefaultWalletSecretKey } from '@app/store/in-memory-key/in-memory-key.selectors';
+import { SecretKey } from '@app/ui/components/secret-key-current/secret-key';
+
+export function UnlockedViewSecretKey() {
+  const defaultWalletSecretKey = useDefaultWalletSecretKey();
+
+  return (
+    <>
+      <Header px="space.04">
+        <HeaderGrid leftCol={<HeaderBackButton />} rightCol={null} />
+      </Header>
+      <Content>
+        <TwoColumnLayout
+          leftColumn={
+            <DescriptionColumn
+              title="Secret Key"
+              description="These 24 words are your Secret Key. They create your account, and you sign in on
+            different devices with them. Make sure to save these somewhere safe. If you lose these
+            words, you lose your account."
+            />
+          }
+          rightColumn={<SecretKey secretKey={defaultWalletSecretKey ?? ''} />}
+        />
+      </Content>
+    </>
+  );
+}
